+++ apps/mobile/src/utils/sanity.ts
@@ -1,5 +1,12 @@
 import { env } from '~/env/runtime';
 // eslint-disable-next-line no-restricted-imports
-import { fetchSanityContent as _fetchSanityContent } from '~/shared/utils/sanity';
+import {
+  fetchSanityContent as _fetchSanityContent,
+  useSanityMaintenanceCheck as _useSanityMaintenanceCheck,
+} from '~/shared/utils/sanity';
 
 export const fetchSanityContent = _fetchSanityContent(env.SANITY_PROJECT_ID);
+
+export const useSanityMaintenanceCheck = () => {
+  return _useSanityMaintenanceCheck(env.SANITY_PROJECT_ID);
+};
